requirements.txt
anthropic>=0.30.0
vertexai>=1.67.1
requests>=2.31.0
PyYAML>=6.0.1
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
pgvector>=0.2.5
orjson>=3.10.7
numpy>=1.26.0
scikit-learn>=1.4.0

